Move charts into folders
